refactor: make new `dhcp_config_utils.c` file
Move `get_config_dhcpinfo` from `src/config.` and
`tests/dhcp/test_dnsmasq.c` (where it was duplicated) into a new file,
`src/dhcp/dhcp_config_utils.c`.
