+++ src/dhcp/CMakeLists.txt
@@ -14,6 +14,16 @@ endif ()
 add_library(dhcp_service dhcp_service.c)
 target_link_libraries(dhcp_service PUBLIC dhcp_config PRIVATE dnsmasq log os)
 
+add_library(dhcp_config_utils dhcp_config_utils.c)
+target_link_libraries(dhcp_config_utils
+  PUBLIC
+    dhcp_config
+  PRIVATE
+    LibUTHash::LibUTHash
+    net
+    os
+)
+
 add_library(dhcp_config INTERFACE)
 set_target_properties(dhcp_config PROPERTIES PUBLIC_HEADER "dhcp_config.h")
 target_link_libraries(dhcp_config INTERFACE LibUTHash::LibUTHash allocs os net)
